app.py
import streamlit as st
from PIL import Image, ImageEnhance, ImageOps
import matplotlib.pyplot as plt
import numpy as np
import cv2
from tensorflow.keras.models import load_model

import os
import requests
import urllib.request


def download_from_google_drive(file_id, dest_path):
    URL = "https://docs.google.com/uc?export=download"

    with requests.Session() as session:
        response = session.get(URL, params={"id": file_id}, stream=True)
        token = get_confirm_token(response)

        if token:
            response = session.get(
                URL, params={"id": file_id, "confirm": token}, stream=True
            )

        save_response_content(response, dest_path)


def get_confirm_token(response):
    for key, value in response.cookies.items():
        if key.startswith("download_warning"):
            return value
    return None


def save_response_content(response, destination, chunk_size=32768):
    with open(destination, "wb") as f:
        for chunk in response.iter_content(chunk_size):
            if chunk:
                f.write(chunk)


@st.cache_resource
def load_trained_model():
    model_path = "xception_model.h5"
    if not os.path.exists(model_path):
        with st.spinner("Downloading model from Google Drive..."):
            file_id = (
                "1Fz2RuXvM54Ym3QCanvMySKME0Y4-GgEb"  # Replace with your actual file ID
            )
            download_from_google_drive(file_id, model_path)
            st.success("Model downloaded.")
    return load_model(model_path)


st.set_page_config(layout="wide", page_title="Image Adjuster")

st.title("Image Augmentation & Analysis App")

st.sidebar.subheader("Image Source")

image_source = st.sidebar.radio("Choose image source:", ["Example", "Upload"])

example_options = {
    "DCIS 1": "images/DCIS_1.png",
    "DCIS 2": "images/DCIS_2.png",
    "Invasive Tumor": "images/Invasive_Tumor.png",
    "Proliferative Invasive Tumor": "images/Prolif_Tumor.png",
}

selected_example = None
uploaded_file = None

if image_source == "Upload":
    uploaded_file = st.sidebar.file_uploader(
        "Upload an image", type=["jpg", "jpeg", "png"]
    )
else:
    selected_example = st.sidebar.selectbox(
        "Select an example image:", list(example_options.keys())
    )

image = None
if image_source == "Upload" and uploaded_file:
    image = Image.open(uploaded_file).convert("RGB")
elif image_source == "Example" and selected_example:
    image_path = example_options[selected_example]
    image = Image.open(image_path).convert("RGB")


model = load_trained_model()


def calculate_image_stats(image):
    """Calculates brightness, saturation, and contrast for an image."""
    image_np = np.array(image.convert("RGB")) / 255.0
    hsv = cv2.cvtColor((image_np * 255).astype(np.uint8), cv2.COLOR_RGB2HSV)

    brightness = np.mean(hsv[:, :, 2])
    saturation = np.mean(hsv[:, :, 1])
    contrast = np.std(
        cv2.cvtColor((image_np * 255).astype(np.uint8), cv2.COLOR_RGB2GRAY)
    )

    return brightness, saturation, contrast


def auto_adjust_image(
    image,
    feature="All",
    brightness_range=(153, 178),
    saturation_range=(52, 76.5),
    contrast_range=(28.4, 43),
):
    brightness, saturation, contrast = calculate_image_stats(image)

    def get_factor(current, range):
        mean = np.mean(range)
        if mean > 0:
            return mean / current
        else:
            return 1.0

    def clamp(f, min_f=0.5, max_f=1.5):
        # return max(min(f, max_f), min_f)
        return f

    brightness_factor = 1.0
    saturation_factor = 1.0
    contrast_factor = 1.0

    if feature in ("Brightness", "All"):
        brightness_factor = clamp(get_factor(brightness, brightness_range))
    if feature in ("Saturation", "All"):
        saturation_factor = clamp(get_factor(saturation, saturation_range))
    if feature in ("Contrast", "All"):
        contrast_factor = clamp(get_factor(contrast, contrast_range))

    image = ImageEnhance.Brightness(image).enhance(brightness_factor)
    image = ImageEnhance.Contrast(image).enhance(contrast_factor)
    image = ImageEnhance.Color(image).enhance(saturation_factor)

    return image


if image:
    # Load and display image
    col1, col2, col3 = st.columns(3)

    with col1:
        st.image(image, caption="Original Image", use_container_width=True)
        st.markdown("**Original Image Stats**")
        orig_bright, orig_sat, orig_con = calculate_image_stats(image)
        st.write(f"Brightness: {orig_bright:.2f}")
        st.write(f"Saturation: {orig_sat:.2f}")
        st.write(f"Contrast: {orig_con:.2f}")

    st.sidebar.subheader("Augmentation Options")

    # Dropdown menu to choose which feature to adjust
    feature_choice = st.sidebar.selectbox(
        "Select a feature to adjust (for both manual and auto adjust):",
        ["Brightness", "Saturation", "Contrast", "All"],
    )

    # Default factors
    brightness_factor = 1.0
    contrast_factor = 1.0
    saturation_factor = 1.0

    # Show slider based on selection
    if feature_choice in ("Brightness", "All"):
        brightness_factor = st.sidebar.slider("Brightness", 0.01, 2.0, 1.0)

    if feature_choice in ("Saturation", "All"):
        saturation_factor = st.sidebar.slider("Saturation", 0.01, 2.0, 1.0)

    if feature_choice in ("Contrast", "All"):
        contrast_factor = st.sidebar.slider("Contrast", 0.01, 2.0, 1.0)

    # Apply augmentations
    augmented = image
    augmented = ImageEnhance.Brightness(augmented).enhance(brightness_factor)
    augmented = ImageEnhance.Contrast(augmented).enhance(contrast_factor)
    augmented = ImageEnhance.Color(augmented).enhance(saturation_factor)

    adjusted = auto_adjust_image(augmented, feature=feature_choice)

    with col2:
        st.image(augmented, caption="Augmented Image", use_container_width=True)
        aug_bright, aug_sat, aug_con = calculate_image_stats(augmented)
        st.markdown("**Augmented Image Stats**")
        st.write(f"Brightness: {aug_bright:.2f}")
        st.write(f"Saturation: {aug_sat:.2f}")
        st.write(f"Contrast: {aug_con:.2f}")

    # Analyze both original and augmented image
    orig_bright, orig_sat, orig_con = calculate_image_stats(image)
    aug_bright, aug_sat, aug_con = calculate_image_stats(augmented)
    adj_bright, adj_sat, adj_con = calculate_image_stats(adjusted)

    with col3:
        st.image(adjusted, caption="Auto Adjusted Image", use_container_width=True)
        st.markdown("**Auto Augmented**")
        st.write(f"Brightness: {adj_bright:.2f}")
        st.write(f"Saturation: {adj_sat:.2f}")
        st.write(f"Contrast: {adj_con:.2f}")

    st.markdown("---")
    # DOESN"T LOAD THE ICON IN MY EDITOR BUT ITS THERE TRUST!"
    st.markdown("### 🔍 Run Prediction")

    with st.form("predict_form"):
        submitted = st.form_submit_button("Predict", use_container_width=True)

        if submitted:
            try:
                if model:
                    st.subheader("🧠 Model Predictions")

                    # Preprocess function
                    def preprocess(img):
                        arr = np.array(img.resize((100, 100))) / 255.0
                        return np.expand_dims(arr, axis=0)

                    # Preprocess all images
                    img_input = preprocess(image)
                    aug_input = preprocess(augmented)
                    adj_input = preprocess(adjusted)

                    # Predictions
                    img_pred = model.predict(img_input)
                    aug_pred = model.predict(aug_input)
                    adj_pred = model.predict(adj_input)

                    img_label = np.argmax(img_pred, axis=1)[0]
                    aug_label = np.argmax(aug_pred, axis=1)[0]
                    adj_label = np.argmax(adj_pred, axis=1)[0]

                    index_to_label = {
                        0: "DCIS_1",
                        1: "DCIS_2",
                        2: "Invasive_Tumor",
                        3: "Prolif_Invasive_Tumor",
                    }

                    pred_cols = st.columns(3)
                    for col, title, label in zip(
                        pred_cols,
                        ["Original", "Augmented", "Auto Adjusted"],
                        [img_label, aug_label, adj_label],
                    ):
                        with col:
                            st.markdown(f"#### {title}")
                            st.success(f"**Prediction:** {index_to_label[label]}")

                else:
                    st.warning("No model found.")
            except Exception as e:
                st.error("Something went wrong during prediction.")
                st.text(str(e))

    accuracy_data = {"Original": 0.85, "Augmented": 0.88, "Auto Adjusted": 0.90}

    st.markdown("---")
    st.markdown("### 📈 Model Accuracy Comparison")

    fig, ax = plt.subplots()
    labels = list(accuracy_data.keys())
    values = list(accuracy_data.values())
    bars = ax.bar(labels, values)
    ax.set_ylim(0, 1)
    ax.set_ylabel("Accuracy")
    ax.set_title("Prediction Accuracy per Image Variant")

    for bar in bars:
        height = bar.get_height()
        ax.annotate(
            f"{height:.2f}",
            xy=(bar.get_x() + bar.get_width() / 2, height),
            xytext=(0, 5),
            textcoords="offset points",
            ha="center",
            va="bottom",
        )

    st.pyplot(fig)
